pyproject.toml
# pyproject.toml
[project]
name = "v0-image-optimizer"
version = "1.0.0"
description = "High-performance CLI for WebP conversion and image optimization"
authors = [{name = "Leonardo Firme", email = "seu-email@exemplo.com"}]
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.2.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "types-pillow>=10.2.0.20240822",
]
